Good practices on > and < operations
Educational note: when histograms are filled, it's typical that a given bin contains entries that are greater than *or equal to* the lower bin boundary and less then (but not equal to!) the upper bin boundary.  I've adjusted the mass window to reflect that (even though admittedly the problem statement was vague on this point).